challenges/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.1
pyjwt==2.8.0
psycopg2-binary==2.9.9
aio_pika==9.4.0
